requirements-dev.txt
pytrutankless==0.1.2
black==22.6.0
isort==5.10.1
pre-commit==2.20.0
pytest-homeassistant-custom-component==0.11.3
yamllint==1.27.1